SystemTypeMapper/flask_backend/requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
numpy
openai
python-dotenv